requirements.txt
flet>=0.24.0
requests>=2.31.0
reportlab>=4.0.0
Pillow>=10.0.0
python-dateutil>=2.8.2
